Рефакторинг структуры Composable компонентов меню
